requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-community
pypdf



#for installation of local packages
-e .